==> Part2_SparsenessCalculation.Rmd <==
---
title: 'Part 2: Sparseness calculation'
author: "Pernille Berg Lassen"
date: "31/3/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---
Calculating sparseness for the three maps types - a saliency map and the three conspicuity maps

The saliency maps must be created in matlab using the GBVS algorithm (Harel, Koch & Perona, 2007)

Equation for calculating sparseness is adapted from Koide et al. (2015)

```{r}
#Load packages - ONLY NECESSARY IF PREVIOUS PARTS HAVE NOT BEEN RUN
library(pacman)
p_load(devtools,ggplot2,tidyverse,lmerTest,data.table,ggpubr,caret,lme4,png,
       dbscan,stringr,gtools,wesanderson,RColorBrewer,matrixStats,topicmodels,
       lattice,ggforce,msm,plyr,dplyr,gplots,groupdata2,hydroGOF)
```


```{r}
##Resized
#Creating an empty dataframe
sparseness_resized_df <- data.frame(ID = numeric(),
                                    L1 = numeric(),
                                    L2 = numeric(),
                                    median_sal = numeric(),
                                    mean_sal = numeric(),
                                    sparseness = numeric())

#Resized is the saliency maps fit to the orignal size of the image - large files, takes a little to load

temp = list.files(path = "screenshots/", pattern="*_resized.csv", full.names = T)
new_temp <- mixedsort(temp) #check that the files are sorted corrected (abs_1-20, and fig_1-20)
myfiles = lapply(new_temp,read.delim,sep=",",header = F)

N_resized = 1764000 #number of pixels in resized images/maps

ID = 1
for (map in myfiles){
  L1 <- sum(map)
  L2 <- sqrt(sum(map^2))
  median_sal_val <- median(as.matrix(map))
  mean_sal_val <- mean(as.matrix(map))
  sparseness_val <- (1/(sqrt(N_resized)-1))*(sqrt(N_resized)-(L1/L2))
  sparseness_resized_df[nrow(sparseness_resized_df) + 1,] = c(ID = ID, L1 = L1, L2 = L2, median_sal = median_sal_val, mean_sal = mean_sal_val, sparseness = sparseness_val)
  ID = ID + 1
}

sparseness_resized_df$artist <- c(rep("MorrisLouis", 6), rep("PietMondrian", 6),
                                  rep("ClaudeMonet", 6), rep("VanGogh", 6)) #Change artists name accordingly

#Conspicuity and 'acutal' maps

N = 576 #number of pixels in conspicuity maps and 'actual' saliency maps

temp = list.files(path = "screenshots/", pattern="*_ori.csv", full.names = T)
new_temp <- mixedsort(temp) #check that the files are sorted corrected (abs_1-20, and fig_1-20)
myfiles = lapply(new_temp,read.delim,sep=",",header = F)

sparseness_ori_df <- data.frame(ID = numeric(), #change name of dataframe accordingly
                                L1 = numeric(),
                                L2 = numeric(),
                                median_sal = numeric(),
                                mean_sal = numeric(),
                                sparseness = numeric())

ID = 1
for (map in myfiles){
  L1 <- sum(map)
  L2 <- sqrt(sum(map^2))
  median_sal_val <- median(as.matrix(map))
  mean_sal_val <- mean(as.matrix(map))
  sparseness_val <- (1/(sqrt(N)-1))*(sqrt(N)-(L1/L2))
  sparseness_ori_df[nrow(sparseness_ori_df) + 1,] = c(ID = ID, L1 = L1, L2 = L2, median_sal = median_sal_val, mean_sal = mean_sal_val, sparseness = sparseness_val)
  ID = ID + 1 #change name of dataframe accordingly
}

sparseness_ori_df$artist <- c(rep("MorrisLouis", 6), rep("PietMondrian", 6),
                                  rep("ClaudeMonet", 6), rep("VanGogh", 6))

#Creating an ID/image coloumn to merge with fixation data
sparseness_resized_df$image <- c(rep(seq(1:12), 2))
sparseness_resized_df$genre <- c(rep('abs',12),rep('fig',12))
sparseness_resized_df$image <- paste(sparseness_resized_df$genre,"_",sparseness_resized_df$image)
sparseness_resized_df$image <- gsub( " ", "", sparseness_resized_df$image) 

#cols.num <- c("L1","L2","median_sal","mean_sal","sparseness") - if non-numeric
#sparseness_df[cols.num] <- sapply(sparseness_df[cols.num],as.numeric)
#Checking classes of variables: sapply(sparseness_df, class)

#sparseness_df %>% mutate_at(vars(L1, L2,median_sal,mean_sal,sparseness), funs(round(., 2))) #Rounding multiple variables - if necessary

median(sparseness_small_df$sparseness) #0.2444834 <- small difference in median from resized but classificaiton is the same between the two
median(sparseness_resized_df$sparseness) # 0.2732241
median(sparseness_color_df$sparseness) #0.2746725
median(sparseness_int_df$sparseness) #0.3061818
median(sparseness_ori_df$sparseness) #0.2725344

sparseness_resized_df$image[sparseness_resized_df$sparseness >= 0.2732241] 
sparseness_small_df$image[sparseness_small_df$sparseness >= 0.2444834] 
sparseness_color_df$image[sparseness_color_df$sparseness >= 0.2746725]
sparseness_int_df$image[sparseness_int_df$sparseness >= 0.3061818]
sparseness_ori_df$image[sparseness_ori_df$sparseness >= 0.2725344]
#Some slight differences in classification between maps

few_peak <- subset(sparseness_resized_df,sparseness_resized_df$sparseness >= 0.2732241)
few_peak$peak_group <- 'few'
many_peak <- subset(sparseness_resized_df,sparseness_resized_df$sparseness < 0.2732241)
many_peak$peak_group <- 'many'

peak <- rbind(few_peak,many_peak)
peak <- subset(peak, select = -c(ID,artist,genre))

write.csv(peak,'SparsenessResults_newStimuli.csv')
```
