==> Part4_PreliminaryDataExp.Rmd <==
---
title: "Part 4: Preliminary data exploration and visualization"
author: "Pernille Berg Lassen"
date: "30/3/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
#Load packages - ONLY NECESSARY IF PREVIOUS PARTS HAVE NOT BEEN RUN
library(pacman)
p_load(devtools,ggplot2,tidyverse,lmerTest,data.table,ggpubr,caret,lme4,png,
       dbscan,stringr,gtools,wesanderson,RColorBrewer,matrixStats,topicmodels,
       lattice,ggforce,msm,plyr,dplyr,gplots,groupdata2,hydroGOF)

#Load data
Fix_df <- read.csv('Fix_data_BSC.csv') #Dataset containing all fixations
Fix_df = subset(Fix_df, select = -c(X))
Fix_df <- Fix_df[order(Fix_df[,2], Fix_df[,18]),] 
#Change coloumn numbers accrodingly

Fix_df_10Fix <- read.csv('Old_data_10Fix.csv') #Dataset contaning the first 10 fixations on each image for each participant
Fix_df_10Fix = subset(Fix_df_10Fix, select = -c(X))
Fix_df_10Fix <- Fix_df[order(Fix_df_10Fix[,2], Fix_df_10Fix[,18]),] 
#Change coloumn numbers accrodingly

Fix_df_fiveS <- read.csv('fixations_first10seconds.csv') #Dataset containing the fixations within the first five seconds on each image for each participant
Fix_df_fiveS = subset(Fix_df_fiveS, select = -c(X))
Fix_df_fiveS <- Fix_df_fiveS[order(Fix_df_fiveS[,2], Fix_df_fiveS[,18]),] #Change coloumn numbers accrodingly

BehaveDat <- read.csv('DataBsC/BehavioralData_New.csv')
BehaveDat <- subset(BehaveDat, select = -c(X))

```

Vizualizations and preliminary data-exploration

```{r}
#Demographics

Dat_grouped = group_by(BehaveDat,ID,gender,condition,age) %>% summarize()

sum(Dat_grouped$gender == 'Female' & Dat_grouped$condition == 1) 
sum(Dat_grouped$gender == 'Female' & Dat_grouped$condition == 0) 
sum(Dat_grouped$gender == 'Male' & Dat_grouped$condition == 1) 
sum(Dat_grouped$gender == 'Male' & Dat_grouped$condition == 0)
round(mean(Dat_grouped$age),2)
round(sd(Dat_grouped$age),2)
round(mean(Dat_grouped$age[Dat_grouped$condition == 1]),2)
round(sd(Dat_grouped$age[Dat_grouped$condition == 1]),2)
round(mean(Dat_grouped$age[Dat_grouped$condition == 0]),2)
round(sd(Dat_grouped$age[Dat_grouped$condition == 0]),2)

#Image familarity 

Image_grouped = group_by(BehaveDat, ID, image, q7, artist, genre) %>% summarize()
img_fam = subset(Image_grouped,Image_grouped$q7 == 'yes')
img_fam = group_by(img_fam,image, artist, genre)  %>% summarize(Number_yes=n())
img_fam$Number_yes = sort(img_fam$Number_yes, decreasing = T)

summary(img_fam$artist) 
summary(img_fam$genre)

sum(img_fam$Number_yes[img_fam$artist == 'chagall']) #change with relevant artist names
sum(img_fam$Number_yes[img_fam$artist == 'schiele'])
sum(img_fam$Number_yes[img_fam$artist == 'rothko'])
sum(img_fam$Number_yes[img_fam$artist == 'pollock'])

sum(img_fam$Number_yes[img_fam$genre == 'fig'])
sum(img_fam$Number_yes[img_fam$genre == 'abstract'])


#Art interest
art_interest <- BehaveDat %>% #summarizing results from art interest questionaire
  dplyr::group_by(ID) %>%
  dplyr::summarize(
    q8 = q8[1],
    q9 = q9[1],
    q10 = q10[1],
    q11 = q11[1],
    condition = condition[1]
  )

#How many participants have taken art-classes?

length(art_interest$ID[art_interest$q11 == "yes" & art_interest$condition == 1])
length(art_interest$ID[art_interest$q11 == "yes" & art_interest$condition == 0])

art_interest$condition = as.factor(art_interest$condition)

art_interest$condition = plyr::revalue(art_interest$condition, c("1"= "Experts", "0"= "Non-Experts"))

all_part <- ggplot(art_interest, aes(x = q8)) +
  geom_bar(fill = "#99CCFF") +
  xlab("Ratings: Question X") +
  ylab("Participant counts") +
  theme_minimal()
  #How to change the values on the y-axis?

#Bar plots
bar_plot <- ggplot(art_interest, aes(x = q8)) + 
  geom_bar(fill = "#99CCFF") + 
  facet_wrap(~condition) + 
  xlab("Ratings: Question X") + 
  ylab("Partcipant counts") + 
  theme_minimal()

box_plot <- ggplot(art_interest,aes(y=q9,x=condition)) +
  geom_boxplot(color = "#336699", fill = "#6699CC") +
  xlab("Condition") +
  ylab("Rating") +
  ggtitle("Question X") + 
  coord_cartesian(ylim = c(1, 7)) +
  theme_minimal()

#Leaving an axis label blank: theme(axis.title.y = element_blank())

ggarrange(all_part,bar_plot) #what plots to display next to one another?


```

Testing art-interest

```{r}
#Mean and SD
art_interest %>%
  dplyr::group_by(condition) %>%
  dplyr::summarize(
    mean_q8 = round(mean(q8),2),
    sd_q8 = round(sd(q8),2),
    mean_q9 = round(mean(q9),2),
    sd_q9 = round(sd(q9),2),
    mean_q10 = round(mean(q10),2),
    sd_q10 = round(sd(q10),2)
  ) #Means are nearly identical - though condition 1 is consistently higher - biggest difference is in question 10

#testing assumptions - what to test?

m.int <- glm(condition ~ q8 + q9 + q10, family = 'binomial', art_interest) #one model per question or one model with all questions?
summary(m.int)
#No significant difference in art interest between the two conditions


```

Scanpaths and heatmaps

```{r}
#Acessing quality of ET data?
#Subsetting dataframe to make heatmaps for 4 chosen images - two abs, two fig (one for each artist)

img <- png::readPNG('DataBsC/heatmaps_scanpaths_images/screenshot_abs12.png')
g <- grid::rasterGrob(img, interpolate = T)

jet.colors <- colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000"))

#One participant, one image
heatmap_P1 <- ggplot(subset(Fix_df, ID==1 & image=='abstract_12'),aes(x = PositionX, y = PositionY)) +
  xlim(0,1600) +
  ylim(0, 900) +
  annotation_custom(g, xmin=-Inf, xmax=Inf, ymin=-0, ymax=900) +
  stat_density2d(geom="raster", aes(fill=..density.., alpha=sqrt(sqrt(..density..))), contour=FALSE, n=1000) +
scale_alpha(range = c(0.1, 0.6)) + scale_fill_gradientn(colours = jet.colors(10), trans='sqrt')

averaged_heatmaps <- Fix_df %>%
  group_by(image) %>%
  mutate(Counter = sequence(rle(ID)$lengths)) %>%
  ungroup() %>%
  group_by(image, Counter) %>%
  summarize(
    PositionX = median(PositionX),
    PositionY = median(PositionY)
  )

#Heatmaps
aveHM <-
  ggplot(subset(averaged_full_data,image == "abstract_12"),aes(x =PositionX, y = PositionY)) +
  xlim(0,1600) +
  ylim(0,900) +
  annotation_custom(g,xmin=-Inf, xmax=Inf, ymin=-0, ymax=900) +
  stat_density2d(geom="raster", aes(fill=..density.., alpha=sqrt(sqrt(..density..))), contour=FALSE, n=1000) + scale_alpha(range = c(0.1, 0.6)) +
  scale_fill_gradientn(colours = jet.colors(10), trans='sqrt') + ggtitle("All Participants")

#Scanpaths
aveSP = ggplot(subset(averaged_full_data, image == "fig_11"), aes(x = PositionX, y = PositionY)) +
  xlim(0,1600) +
  ylim(0,900) +
  annotation_custom(g, xmin=-Inf, xmax=Inf, ymin=-0, ymax=900) +
  geom_point(size = 8, alpha = 0.5, color = "white") + 
  geom_path(size = 1, alpha = 0.3) + 
  geom_text(aes(label = Counter, size = 5)) +
  ggtitle("All Participants")

#One participant
test_fix <- Fix_df %>%
  group_by(image) %>%
  mutate(Counter = sequence(rle(ID)$lengths))

SP = ggplot(subset(test_fix, image == "abstract_12" & ID == 2), aes(x = PositionX, y = PositionY)) +
  xlim(0,1600) +
  ylim(0,900) +
  annotation_custom(g, xmin=-Inf, xmax=Inf, ymin=-0, ymax=900) +
  geom_point(size = 8, alpha = 0.5, color = "white") + 
  geom_path(size = 1, alpha = 0.3) + 
  geom_text(aes(label = Counter, size = 5)) +
  ggtitle("Participant X")

```

Testing rating scores (Behavioral responses)

```{r}
#Check distribution?

small_BehaveDat <- BehaveDat %>% #(rounded) means for each participant
  dplyr::group_by(ID) %>%
  dplyr::summarize(
    q1 = q1[1],
    q2 = q2[1],
    q3 = q3[1],
    q4 = q4[1],
    q5 = q5[1],
    q6 = q6[1],
    q7 = q7[1],
    q8 = q8[1],
    q9 = q9[1],
    q10 = q10[1],
    q11 = q11[1],
    condition = condition[1]
  )

small_BehaveDat$condition <- as.factor(small_BehaveDat$condition)

#Calculating means for understanding and liking
BehaveDat$LikingMean = round((BehaveDat$q1 + BehaveDat$q2 + BehaveDat$q3)/3, digits = 2)
BehaveDat$UnderstandingMean = round((BehaveDat$q4 + BehaveDat$q5 + BehaveDat$q6)/3, digits = 2)

#Boxplots
ggplot(small_BehaveDat,aes(y=q1,x=condition))+geom_boxplot()+ggtitle("Go to a museum to see this picture?")+
  xlab("Condition") + ylab("Rating")+ coord_cartesian(ylim = c(1, 7))


mean_BehaveDat <- BehaveDat %>% #summarizing mean of ratings for each participants 
  dplyr::group_by(ID) %>%
  dplyr::summarize(
    condition = condition[1],
    q1 = q1[1],
    q2 = q2[1],
    q3 = q3[1],
    q4 = q4[1],
    q5 = q5[1],
    q6 = q6[1],
    LikingMean = LikingMean[1],
    UnderstandingMean = UnderstandingMean[1]
  )

BehaveDat$condition <- as.factor(BehaveDat$condition)

rescalelist = c("LikingMean","UnderstandingMean") #making a list with variales we want to scale
BehaveDat.s = BehaveDat[, colnames(BehaveDat) %in% rescalelist] %>% 
  apply(.,function(x) scale(x,center= mean(x,na.rm = T), 
                            scale = sd(x, na.rm = T))) %>% 
  cbind(.,BehaveDat[,! colnames(BehaveDat) %in% rescalelist]) 


m1.1 = lmerTest::lmer(LikingMean ~ 1 +condition + (1|ID) +
                        (1+condition|image),BehaveDat)

m1.2 = lmerTest::lmer(UnderstandingMean ~ 1 + condition +
                        (1|ID)+(1+condition|image),BehaveDat) 
#A significant effect of condition

m1.3 = lmerTest::lmer(LikingMean ~ 1 + condition*artist +
                        (1+artist|ID)+(1+condition|image),BehaveDat)

m1.4 = lmerTest::lmer(UnderstandingMean ~ 1 + condition*artist+
                        (1+artist|ID)+(1+condition|image), BehaveDat) 
#A significant effect of rothko and pollock. Also a significant interaction effect between pollock and condition

m1.5 <- lmerTest::lmer(LikingMean ~ 1 + condition*genre+
                        (1+genre|ID)+(1+condition|image), BehaveDat) 

summary(m1.5)

m1.6 <- lmerTest::lmer(UnderstandingMean ~ 1 + condition*genre+
                        (1+genre|ID)+(1+condition|image), BehaveDat) 

summary(m1.6)

#Defing a function to get performance
getPerformance = function(test_df, train_df, mdl, mdl_string, n = NA){
  #asses performance and returns a result df
  
    #save perf to list
      #Test performance
  
  #extract predicted value from the mdl string to use in the rmse
  temp_string = gsub("(\\~).+", mdl_string, replacement = "")
  actual_col = gsub(" ", x = temp_string, replacement = "")
  actual =pull(dplyr::select(test_df, actual_col))
  #calculating rmse
  rmse = hydroGOF::rmse(predict(mdl, test_df, allow.new.levels = T), actual , na.rm = T)
  mdlPerf = summary(mdl)
    #saving performance metrix to a df
  result_df =  data.frame(rmse = rmse,
                          AIC = mdlPerf$AICtab[1],
                          n = n) 
  return(result_df)
} 

#defining a cross validate function
CrossVal = function(num_folds, dataset, mdl_string, ID_col = NULL, CAT_col = NULL) {
  
  #folding the dataset
  dataset = fold(dataset, num_folds, cat_col = CAT_col, id_col = ID_col, method = 'n_dist')
  
  #looping through the folds
  for (fold in seq(num_folds)) {
    train_df = subset(dataset, .folds != fold)
    test_df = subset(dataset, .folds == fold)
    

    mdl = try(lmer(mdl_string, train_df, 
                   control = lmerControl(optimizer = "nloptwrap", calc.derivs = FALSE)))
    temp_sum = try(summary(mdl))
    if (length(temp_sum) > 3){ #if you could make a model
      #asses performance and append it to a df
      temp = getPerformance(test_df, train_df, mdl, mdl_string, n = fold)
    } else {#if you couldn't make a model
      temp = data.frame(rmse = NA,
                        AIC = NA,
                        n = n)
    }
    temp$mdl = mdl_string
    temp$numfolds = num_folds
    if (fold == 1){ #if first part - make a df
      perf_df = temp
    } else { #else append to df
      perf_df = rbind(perf_df, temp)  
    }
    
  }
  return(perf_df)
}

BehaveDat$ID <- as.factor(BehaveDat$ID)

CV_m1.1 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "LikingMean ~ 1 +condition + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

CV_m1.2 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "UnderstandingMean ~ 1 +condition + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

CV_m1.3 <- CV_m1.1 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "LikingMean ~ 1 +condition*artist + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

CV_m1.4 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "UnderstandingMean ~ 1 +condition*artist + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

CV_m1.5 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "LikingMean ~ 1 +condition*genre + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

CV_m1.6 <- CrossVal(num_folds = 3, dataset = BehaveDat, mdl_string = "UnderstandingMean ~ 1 +condition*genre + (1|ID) + (1+condition|image)", ID_col = "ID", CAT_col = 'condition')

perf_df_behave = rbind(CV_m1.1,CV_m1.2,CV_m1.3,CV_m1.4,CV_m1.5,CV_m1.6)
perf_df_sum_behave = group_by(perf_df_behave, mdl) %>%
  summarise_all(funs(mean(., na.rm = TRUE)))

perf_df_sum_behave <- perf_df_sum_behave[order(perf_df_sum_behave$rmse),]

```

Discrimination task

```{r}

#accuracy is a binary variable - 0 or 1

#how to best visualize?

hi$accuracy <- as.factor(hi$accuracy)

m3.1 <- glmer(accuracy ~ 1 + condition + 
                (1|ID) + (1|image), data = hi, family = 'binomial')

m3.2 <- glmer(accuracy ~ 1 + condition*artist + 
                + (1|image), data = hi, family = 'binomial')

m3.3 <- glmer(accuracy ~ 1 + condition*genre + 
                (1|ID) + (1|image), data = hi, family = 'binomial')

```
