==> Part6_TestingAUCscores.Rmd <==
---
title: 'Part 6: Testing AUC scores'
author: "Pernille Berg Lassen"
date: "31/3/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
#Load packages - ONLY IF NOT RUN IN A PREVIOUS PART
library(pacman)
p_load(devtools,ggplot2,tidyverse,lmerTest,data.table,ggpubr,caret,lme4,png,
       dbscan,stringr,gtools,wesanderson,RColorBrewer,matrixStats,topicmodels,
       lattice,ggforce,msm,plyr,dplyr,gplots,groupdata2,hydroGOF)
```

```{r}
#Reading in AUC scores - full set, first 10 fixations
auc_df <- read.csv('AUC_scores_10Fix.csv')
auc_df <- merge(auc_df,sparsenessResults)
auc_df <- subset(auc_df, select = -c(X))
auc_df <- auc_df[order(auc_df[,1], auc_df[,2],auc_df[,4]),]


#Reading in AUC scores - five first seconds (missing P1 and P13)
auc_5s <- read.csv('AUC_scores_5sec.csv')
auc_5s <- merge(auc_5s,sparsenessResults)
auc_5s <- subset(auc_5s, select = -c(X))
auc_5s <- auc_5s[order(auc_5s[,1], auc_5s[,2],auc_5s[,3]),]

```

```{r}
#Adding artist:
pollock_list <- c('abstract_11','abstract_12','abstract_13','abstract_14','abstract_15','abstract_16','abstract_17','abstract_18','abstract_19','abstract_20')

rothko_list <- c('abstract_1','abstract_2','abstract_3','abstract_4','abstract_5','abstract_6','abstract_7','abstract_8','abstract_9','abstract_10')

chagall_list <- c('fig_1','fig_2','fig_3','4','fig_5','fig_6','fig_7','fig_8','fig_9','fig_10')

schiele_list <- c('fig_11','fig_12','fig_13','fig_14','fig_15','fig_16','fig_17','fig_18','fig_19','fig_20')

auc_df$artist <- ifelse(auc_df$image %in% pollock_list, 'pollock',NA)

auc_df <- mutate(auc_df,
       artist = case_when(auc_df$image %in% pollock_list ~ "pollock",
                          auc_df$image %in% rothko_list ~ 'rothko',
                          auc_df$image %in% schiele_list ~ 'schiele',
                          auc_df$image %in% chagall_list ~ 'chagall'
       ))

auc_5s <- mutate(auc_5s,
       artist = case_when(auc_5s$image %in% pollock_list ~ "pollock",
                          auc_5s$image %in% rothko_list ~ 'rothko',
                          auc_5s$image %in% schiele_list ~ 'schiele',
                          auc_5s$image %in% chagall_list ~ 'chagall'
       ))
```

```{r}
#These AUC scores was calculated from the 10 first fixations for each participant for each image
auc_df$condition <- as.factor(auc_df$condition)
auc_df$condition=plyr::revalue(auc_df$condition,c("0"="Non-experts","1"="Experts"))
auc_df$genre <- str_replace_all(auc_df$genre,"_","")

auc_5s$condition <- as.factor(auc_5s$condition)

#Plots

plot_data <- auc_df %>%
  dplyr::group_by(condition, map_type, peak_group) %>%
  dplyr::summarize(
    mean = mean(AUC),
    sd = sd(AUC),
    se = sd(AUC)/sqrt(length(AUC))
  )

pd <- position_dodge(0.5)

ggplot(plot_data, aes(x=condition, y=mean, colour=map_type)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_line(position = pd) +
    geom_point(position = pd) +
    facet_wrap(~peak_group) +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Salience effects for the first 10 fixations') +
    ylim(0.75,1) +
    scale_color_manual(values = wes_palette(n = 4, name ="GrandBudapest1"))

#Barplot - only for one type of plot
ggplot(subset(plot_data, plot_data$map_type == "combined"), aes(x = condition, y = mean, fill = condition)) + 
  geom_col() + 
  geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1) +
  facet_wrap(~peak_group) +
  scale_fill_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
  ggtitle('Salience effects for the first 10 fixations') +
  xlab('Condition') +
  ylab('mean salience effects (AUCs)') +
  guides(fill = "none")

#Over fixation index

auc_df$Fixation_Index <- as.numeric(auc_df$Fixation_Index)

#Full data set
plot_data <- subset(auc_df, auc_df$map_type == 'combined') %>%
  dplyr::group_by(Fixation_Index,condition, genre,map_type, peak_group) %>%
  dplyr::summarize(
    mean = mean(AUC),
    sd = sd(AUC),
    se = sd(AUC) / sqrt(length(AUC))
  )

plot_data$Fixation_Index <- as.character(plot_data$Fixation_Index)
#test_full$condition <- as.character(test_full$condition)

ggplot(subset(plot_data, plot_data$peak_group == 'few'), aes(x=Fixation_Index, y=mean, color = condition)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_point(position = pd) +
    geom_line(aes(group = condition), position = pd) +
    xlab('fixation index') +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Variation of salience effects over fixations') +
    ylim(0.7,1)+
    scale_x_discrete(limits=c("1","2","3","4","5","6","7","8","9","10"))+
    scale_color_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
    facet_wrap(~genre)

ggplot(subset(plot_data, plot_data$peak_group == 'many'), aes(x=Fixation_Index, y=mean, color = condition)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_point(position = pd) +
    geom_line(aes(group = condition), position = pd) +
    xlab('fixation index') +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Variation of salience effects over fixations') +
    ylim(0.65,1)+
    scale_x_discrete(limits=c("1","2","3","4","5","6","7","8","9","10"))+
    scale_color_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
    facet_wrap(~genre)


#Full data set - first five seconds
plot_data <- auc_5s %>%
  dplyr::group_by(condition, map_type,peak_group) %>%
  dplyr::summarize(
    mean = mean(AUC),
    sd = sd(AUC),
    se = sd(AUC) / sqrt(length(AUC)),
    lower_bound = mean - se,
    upper_bound = mean + se
  )

auc_5s$condition=plyr::revalue(auc_5s$condition,c("0"="Non-experts","1"="Experts"))

ggplot(plot_data, aes(x=condition, y=mean, colour=map_type)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_line(position = pd) +
    geom_point(position = pd) +
    facet_wrap(~peak_group) +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Salience effects for the first five seconds') +
    ylim(0.75,1) +
    scale_color_manual(values = wes_palette(n = 4, name ="GrandBudapest1"))

#Barplot - only for one type of map
ggplot(subset(plot_data, plot_data$map_type == "combined"), aes(x = condition, y = mean, fill = condition)) + 
  geom_col() + 
  geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1) +
  facet_wrap(~peak_group) +
  scale_fill_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
  ggtitle('Salience effects for the first 10 fixations') +
  xlab('Condition') +
  ylab('mean salience effects (AUCs)') +
  guides(fill = "none")


plot_data <- subset(auc_5s, auc_5s$map_type == 'combined') %>%
  dplyr::group_by(Fixation_Index,condition, genre,map_type, peak_group) %>%
  dplyr::summarize(
    mean = mean(AUC),
    sd = sd(AUC),
    se = sd(AUC) / sqrt(length(AUC))
  )


plot_data$Fixation_Index <- as.character(plot_data$Fixation_Index)
#test_full$condition <- as.character(test_full$condition)

plot_data <- na.omit(plot_data)
plot_data <- subset(plot_data, plot_data$Fixation_Index != '14')

ggplot(subset(plot_data, plot_data$peak_group == 'few'), aes(x=Fixation_Index, y=mean, color = condition)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_point(position = pd) +
    geom_line(aes(group = condition), position = pd) +
    xlab('fixation index') +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Variation of salience effects over fixations (Few-peaked group)') +
    ylim(0.7,1)+
    scale_x_discrete(limits=c("1","2","3","4","5","6","7","8","9","10"))+
    scale_color_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
    facet_wrap(~genre)

ggplot(subset(plot_data, plot_data$peak_group == 'many'), aes(x=Fixation_Index, y=mean, color = condition)) + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se), width=.1, position = pd) +
    geom_point(position = pd) +
    geom_line(aes(group = condition), position = pd) +
    xlab('fixation index') +
    ylab('mean salience effects (AUCs)') +
    ggtitle('Variation of salience effects over fixations (Many-peaked group') +
    ylim(0.65,1)+
    scale_x_discrete(limits=c("1","2","3","4","5","6","7","8","9","10"))+
    scale_color_manual(values = wes_palette(n = 2, name = "GrandBudapest1")) +
    facet_wrap(~genre)

#stat_summary(fun = mean,geom="line",aes(group=1))

```

Models

```{r}

#Scale variables?
#rescalelist = c("AUC") #making a list with variales we want to scale
#auc_df.s = auc_df[, colnames(auc_df) %in% rescalelist] %>% 
#  lapply(.,function(x) scale(x,center= mean(x,na.rm = T), 
#                             scale = sd(x, na.rm = T))) %>% 
#  cbind(.,auc_df[,! colnames(auc_df) %in% rescalelist]) 

#AUC_data_full$ID <- as.integer(AUC_data_full$ID)
#AUC_data_full$genre <- as.factor(AUC_data_full$genre)
#AUC_data_full$Fixation_Index <- factor(AUC_data_full$Fixation_Index, ordered = T)

m1 <- lmer(AUC ~ 1 + condition*Fixation_Index*peak_group + 
             (1|ID) + 
             (1+condition|image),data = auc_5s,
           control=lmerControl(optimizer = "nloptwrap",calc.derivs = F))
###Troubles with this model - doesn't run
#summary(m1)

#m2 <- lmer(AUC ~ 1 + condition*Fixation_Index*peak_group*map_type + (1|ID) + (1|image), data = auc_df)
#summary(m2)

m3 <- lmer(AUC ~ 1 + condition*Fixation_Index*peak_group*map_type + (1|ID) + (1+condition|image), data = auc_5s)
summary(m3)

#m4 <- lmer(AUC ~  1 + condition*Fixation_Index*peak_group*map_type*artist + (1|ID) + (1|image), data = auc_df)

m5 <- lmer(AUC ~  1 + condition*Fixation_Index*peak_group*map_type*artist + (1|ID) + (1|image), data = auc_5s)
summary(m5)

#m6 <- lmer(AUC ~  1 + condition*Fixation_Index*peak_group*map_type*genre + (1|ID) + (1|image), data = auc_df)

m7 <- lmer(AUC ~  1 + condition*Fixation_Index*peak_group*map_type*genre + (1|ID) + (1|image), data = auc_5s)
summary(m7)
```

```{r}
#Defing a function to get performance
getPerformance = function(test_df, train_df, mdl, mdl_string, n = NA){
  #asses performance and returns a result df
  
    #save perf to list
      #Test performance
  
  #extract predicted value from the mdl string to use in the rmse
  temp_string = gsub("(\\~).+", mdl_string, replacement = "")
  actual_col = gsub(" ", x = temp_string, replacement = "")
  actual =pull(dplyr::select(test_df, actual_col))
  #calculating rmse
  rmse = hydroGOF::rmse(predict(mdl, test_df, allow.new.levels = T), actual , na.rm = T)
  mdlPerf = summary(mdl)
    #saving performance metrix to a df
  result_df =  data.frame(rmse = rmse,
                          AIC = mdlPerf$AICtab[1],
                          n = n) 
  return(result_df)
} 

#defining a cross validate function
CrossVal = function(num_folds, dataset, mdl_string, ID_col = NULL, CAT_col = NULL) {
  
  #folding the dataset
  dataset = fold(dataset, num_folds, cat_col = CAT_col, id_col = ID_col, method = 'n_dist')
  
  #looping through the folds
  for (fold in seq(num_folds)) {
    train_df = subset(dataset, .folds != fold)
    test_df = subset(dataset, .folds == fold)
    

    mdl = try(lmer(mdl_string, train_df, 
                   control = lmerControl(optimizer = "nloptwrap", calc.derivs = FALSE)))
    temp_sum = try(summary(mdl))
    if (length(temp_sum) > 3){ #if you could make a model
      #asses performance and append it to a df
      temp = getPerformance(test_df, train_df, mdl, mdl_string, n = fold)
    } else {#if you couldn't make a model
      temp = data.frame(rmse = NA,
                        AIC = NA,
                        n = n)
    }
    temp$mdl = mdl_string
    temp$numfolds = num_folds
    if (fold == 1){ #if first part - make a df
      perf_df = temp
    } else { #else append to df
      perf_df = rbind(perf_df, temp)  
    }
    
  }
  return(perf_df)
}

auc_5s$ID <- as.factor(auc_5s$ID)

CV_m3 =  CrossVal(num_folds = 3, dataset = auc_5s, mdl_string = "AUC ~ 1 + condition*Fixation_Index*peak_group*map_type + (1|ID) + (1|image)", ID_col = "ID", CAT_col = 'condition')

CV_m5 =  CrossVal(num_folds = 3, dataset = auc_5s, mdl_string = "AUC ~ 1 + condition*Fixation_Index*peak_group*map_type*artist + (1|ID) + (1|image)", ID_col = "ID", CAT_col = 'condition')

CV_m7 =  CrossVal(num_folds = 3, dataset = auc_5s, mdl_string = "AUC ~ 1 + condition*Fixation_Index*peak_group*map_type*genre + (1|ID) + (1|image)", ID_col = "ID", CAT_col = 'condition')

perf_df = rbind(CV_m3,CV_m5, CV_m7)
perf_df_sum = group_by(perf_df, mdl) %>%
  summarise_all(funs(mean(., na.rm = TRUE)))

perf_df_sum <- perf_df_sum[order(perf_df_sum$rmse),]

```
